requirements.txt
deepdiff==5.8.1
GitPython==3.1.27
PyGithub==1.55
PyYAML==6.0
docker==6.0.1
allure-pytest==2.12.0
tavern==1.23.1
pytest==7.1.2
jinja2==3.1.2
semver==2.13.0